+++ src/psync/poverlay.c
@@ -8,6 +8,8 @@
  * the LICENSE file that was distributed with this source code.
  */
 
+#include "pcloudcc/psync/compat.h"
+
 #include "plibs.h"
 #include "poverlay.h"
 #include "ppathstatus.h"
@@ -16,16 +18,14 @@
 int overlays_running = 1;
 int callbacks_running = 1;
 
-#if defined(P_OS_WINDOWS)
+#ifdef P_OS_WINDOWS
 #include "poverlay-windows.c"
-#elif defined(P_OS_LINUX)
-#include "poverlay-linux.c"
-#elif defined(P_OS_MACOSX) || defined(P_OS_BSD)
-#include "poverlay-macos.c"
+#elif defined P_OS_POSIX
+#include "poverlay-posix.c"
 #else
 void overlay_main_loop(void) {}
 void instance_thread(void* payload) {}
-#endif  /* P_OS_WINDOWS */
+#endif /* P_OS_WINDOWS */
 
 poverlay_callback * callbacks;
 static int callbacks_size = 15;
@@ -50,8 +50,8 @@ int psync_add_overlay_callback(int id, poverlay_callback callback) {
 }
 
 void init_overlay_callbacks() {
-  callbacks = (poverlay_callback *) psync_malloc(sizeof(poverlay_callback)*callbacks_size);
-  memset(callbacks, 0, sizeof(poverlay_callback)*callbacks_size);
+  callbacks = (poverlay_callback *) psync_malloc(sizeof(poverlay_callback) * callbacks_size);
+  memset(callbacks, 0, sizeof(poverlay_callback) * callbacks_size);
 }
 
 void psync_stop_overlays() {
@@ -70,58 +70,64 @@ void psync_start_overlay_callbacks() {
   callbacks_running = 1;
 }
 
-void get_answer_to_request(message *request, message *replay) {
-  psync_path_status_t stat=PSYNC_PATH_STATUS_NOT_OURS;
-  memcpy(replay->value, "Ok.", 4);
-  replay->length=sizeof(message)+4;
+void get_answer_to_request(message *request, message *response) {
+  psync_path_status_t stat = PSYNC_PATH_STATUS_NOT_OURS;
+  memcpy(response->value, "Ok.", 4);
+  response->length = sizeof(message) + 4;
+  int max_band;
 
   if (request->type < 20 ) {
     if (overlays_running)
-      stat=psync_path_status_get(request->value);
+      stat = psync_path_status_get(request->value);
     switch (psync_path_status_get_status(stat)) {
       case PSYNC_PATH_STATUS_IN_SYNC:
-        replay->type=10;
+        response->type = 10;
         break;
       case PSYNC_PATH_STATUS_IN_PROG:
-        replay->type=12;
+        response->type = 12;
         break;
       case PSYNC_PATH_STATUS_PAUSED:
       case PSYNC_PATH_STATUS_REMOTE_FULL:
       case PSYNC_PATH_STATUS_LOCAL_FULL:
-        replay->type=11;
+        response->type = 11;
         break;
       default:
-        replay->type=13;
-        memcpy(replay->value, "No.", 4);
+        response->type = 13;
+        memcpy(response->value, "No.", 4);
     }
-  } else if ((callbacks_running)&&(request->type < (callbacks_lower_band + callbacks_size))) {
-    int ind = request->type - 20;
-    int ret = 0;
+
+    return; /* exit */
+  }
+
+  max_band = callbacks_lower_band + callbacks_size;
+  if (psync_overlays_running() && (request->type < max_band)) {
+    uint32_t ind = request->type - 20;
+    int ret;
     message *rep = NULL;
 
     if (callbacks[ind]) {
       if ((ret = callbacks[ind](request->value, rep)) == 0) {
         if (rep) {
-          psync_free(replay);
-          replay = rep;
-        }
-        else
-        replay->type = 0;
+          psync_free(response);
+          response = rep;
+        } else
+          response->type = 0;
       } else {
-        replay->type = ret;
-        memcpy(replay->value, "No.", 4);
+        response->type = ret;
+        memcpy(response->value, "No.", 4);
       }
     } else {
-      replay->type = 13;
-      memcpy(replay->value, "No callback with this id registered.", 37);
-      replay->length = sizeof(message)+37;
-    }
-  } else {
-      replay->type = 13;
-      memcpy(replay->value, "Invalid type.", 14);
-      replay->length = sizeof(message)+14;
+      response->type = 13;
+      memcpy(response->value, "No callback with this id registered.", 37);
+      response->length = sizeof(message)+37;
     }
 
+    return;  /* exit */
+  }
+
+  response->type = 13;
+  memcpy(response->value, "Invalid type.", 14);
+  response->length = sizeof(message) + 14;
 }
 
 int psync_overlays_running() {